requirements.txt
numpy
pandas
torch==1.9.0+cu111
tqdm
wandb
dataclasses
requests
omegaconf
pytorch_lightning==1.4.8
rdkit-pypi
scikit_learn
